Helper for Displaying HTTP Errors
A helper function named DisplayAppError should be written to provide error messages in
JSON as the HTTP response. Client applications can check the HTTP status code to verify
whether the HTTP request is successful. A struct type named appError should be used to
create the model object for providing error messages.

In the appError type, the Error property is used for holding the string value of the error object,
the Message property is used for holding a custom message on the error, and the HttpStatus
property is used for holding HTTP status code. An instance of the errorResource type is
created by providing the value of appError to encode the response as JSON.

Closes #1